CC-14716 Restore correct code

+++ src/SprykerEco/Zed/Unzer/Business/Payment/Processor/PreparePayment/UnzerPreparePaymentProcessor.php
@@ -61,10 +61,10 @@ class UnzerPreparePaymentProcessor implements UnzerPreparePaymentProcessorInterf
         $unzerPaymentTransfer = $quoteTransfer->getPaymentOrFail()->getUnzerPaymentOrFail();
 
         $unzerBasket = $this->createUnzerBasket($quoteTransfer, $unzerPaymentTransfer->getUnzerKeypairOrFail());
-        $amountTotal = (int)$quoteTransfer->getTotalsOrFail()->getGrandTotal() / UnzerConstants::INT_TO_FLOAT_DIVIDER;
+        $amountTotal = (float)$quoteTransfer->getTotalsOrFail()->getGrandTotal() / UnzerConstants::INT_TO_FLOAT_DIVIDER;
         $unzerPaymentTransfer->setBasket($unzerBasket);
         $unzerPaymentTransfer->setCurrency($quoteTransfer->getCurrencyOrFail()->getCode());
-        $unzerPaymentTransfer->setAmountTotal((int)$amountTotal);
+        $unzerPaymentTransfer->setAmountTotal($amountTotal);
 
         return $unzerPaymentTransfer;
     }
